// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/appointments/shared/dto/patient-visit-list.dto.ts
export class GetPatientVisits_DTO {
  PatientId: number = 0;
  PatientCode: string = "";
  FirstName: string = "";
  MiddleName: string = "";
  LastName: string = "";
  ShortName: string = "";
  DateOfBirth: string = "";
  Age: number = 0;
  PhoneNumber: string = "";
  Gender: string = "";
  CountryId: number = 0;
  CountrySubDivisionId: number = 0;
  MunicipalityId: number = 0;
  WardNumber: number = 0;
  CountrySubDivisionName: string = "";
  MunicipalityName: string = "";
  PANNumber: string = "";
  Address: string = "";
  IDCardNumber: string = "";
  DependentId: string = "";
  Posting: string = "";
  Rank: string = "";
  PatientVisitId: number = 0;
  VisitCode: string = "";
  ParentVisitId: number = 0;
  PerformerId: number = 0;
  PerformerName: string = "";
  ReferredById: number = 0;
  VisitDate: string = "";
  AdmittingDate: string = "";
  DischargeDate: string = "";
  VisitTime: string = "";
  DepartmentId: number = 0;
  Department: string = "";
  VisitType: string = "";
  AppointmentType: string = "";
  DaysPassed: number = 0;
  MaxAllowedFollowUpDays: number = 0;
  MaxAllowedReferralDays: number = 0;
  SchemeId: number = 0;
  SchemeName: string = "";
  QueueNo: number = 0;
  IsVisitContinued: boolean = false;
  BillingStatus: string = "";
  PolicyNo: string = "";
  PriceCategoryId: number = 0;
  IsFreeVisit: boolean = false;
  CareTakerName: string = "";
  RelationWithCareTaker: string = "";
  CareTakerContact: string = "";
  MaritalStatus: string = "";
  EthnicGroup: string = "";
}